+++ src/assets/json/translations/de_DE.json
@@ -57,6 +57,7 @@
                 "disabled": "{success.emote} Erfolgreich Antispam deaktiviert."
             },
             "antiinvite": {
+                "enabled": "{success.emote} Erfolgreich Anti-Einladungen mit der Aktion `%c` aktiviert.",
                 "disabled": "{success.emote} Anti-invite erfolgreich deaktiviert."
             },
             "antilink": {
